# Add fast paths for object Get/Set for simple objects