## Jint.Benchmark.ClassBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |ConstructSimple|139.6 ms|2.40 ms|369.26 MB|
| **New** |	| **134.8 ms (-3%)** | **2.11 ms** | **369.26 MB (0%)** |
| Old |ConstructDeepInheritance|157.5 ms|3.15 ms|330.2 MB|
| **New** |	| **152.1 ms (-3%)** | **3.00 ms** | **330.2 MB (0%)** |
| Old |GetSet|120.1 ms|2.00 ms|343.33 MB|
| **New** |	| **117.8 ms (-2%)** | **1.20 ms** | **343.33 MB (0%)** |